Made adjustements to the cradle.
With the Rev A PCBs in hand, I discovered that the screw heads holding
the cradle would interfere with the power and motor connectors, so I
moved them inward.

I also extended the "index" pin on the cradle up through the base plate
to ensure proper orientation of the PCB.

The cradle is too flimsy.  I plan to address that soon.