Experimenting with a better cradle and a limit switch mounting block.
The cradle was twising easily, and, since it's only screwed in from
one side, the other long edge was drooping.  This is an attempt to
beef-up the cradle to prevent that.

I got feedback that having the limit switch soldered directly to the
PCB might cause strain on the PCB that could reduce its lifetime, so I
also started designing a mounting block to hold the limit switch.  The
idea is to make the precise position adjustable with screws.