Fixed a bug and experimented with a mounting block for the limit switch.
The bug occurred when the width of the groove in the spool became
negative, causing the polygon profile of the spool to have a whole,
which prevented OpenSCAD from generating a valid mesh.

I considered removing the limit switch from the PCB over concern that
the repeated action might stress the PCB, shortening it's life.  I
started designing a way to mount the switch (in a different orientation)
to detect the gap in the drive gear teeth.  The idea was the mount the
switch to a block that could be positioned along some rails at assembly
time to ensure it was in exactly the right position.  This got overly
complicated and wasn't going well.  Given that, and that I'm not
convinced the stress of the switch on the PCB is significant, I've put
the switch mount behind a parameter and am putting that on the back
burner for now.