Updated PCB layout (now Rev. B) for more clearance around connectors.
The motion sensor header was within the margin, but the motion sensor
itself would run into the edge of the deeper cradle.  I centered it
on the short axis and aligned it with the index post on the long axis.
This should enable the cradle to have some supports for the sensor
module or a cable that connects to it.