Added current-limiting resistor to base of Q2.
With the thermal camera, I happened to notice that Q1 was heating up
more than I expected.  When on, it was passing almost 11mA.  Within
spec, but way more than necessary to turn on Q2.  Added a current
limited resistor to the base of Q2 dropped the current (and power)
through Q1 to less than 1mA.

Made additional layout tweaks to PCB and updated markings to Rev C.