Redesign of slightly smarter PCB.
Simplified by using a MOSFET, which requires fewer parts and runs the
motor more efficiently.  Since I'm also working on a re-do of the
printed parts, I went ahead and changed the layout and mounting options
for the board.  That makes the new PCB incompatible with the old
mechanical design.