Added a cap screw to hold the drive, and adjusted the new axle.
The cap screw kinda works when printing with 0.3mm layers, but it
would work better at 0.2mm.  It's essentially a 1/4"x20 bolt.  In a
pinch you can force in a tripod mounting screw.

I'd gotten to a nice fit with a prototype axle, but when it was printed
as part of the base plate, the shrinkage was different and thus the
spool assembly didn't fit at all.  After more experimentation, I dialed
in a reasonable default for the "fudge" value to control the tightness.

I'm still not convinced it'll work for everyone given differences in
machines, slicers, materials, etc.  I'm considering making it a
separate part that attaches to the base plate, so you could re-print it
with a different fudge value if it doesn't fit (or if it snaps off).
I might also experiment with something closer to "grip fins" as the
Slant3D guy calls it.