Another fix to shaft adapters (and cap screw).
Recent changes prevented the shaft adapter from sliding all the way
down to the JGY motor base, which meant the drive gear was too high
and wouldn't depress the switch on the Slightly Smarter PCB.

Fix was to make the cap screw diameter larger than the shaft diameter.
To get good threads, the adapter and the cap screw require 0.2mm layers
(or better).  The rest of the parts can still be printed with 0.3mm.

Now it's almost too close, as the switch lever drags in the channel on
the drive gear.  I can look at that later.

This commit also includes a little code cleanup.