Made more tweaks to the build plate.
Added a pocket to hold the end of the PCB near the power connector.  It
has a drooping bridge and needs some sort of support.

Turned off the spool guard experiment.  It doesn't help.

Tried again to make a tighter fitting axle.  This works better, but the
variation from print-to-print means it won't always be better.